requirements.txt
pandas==1.5.1
requests==2.28.1
lxml==4.9.2
openpyxl==3.0.9
pyyaml==6.0
matplotlib==3.5.2
mplfinance==v0.12.9b1
pybit==2.4.1
autopep8==2.0.1
pylint==2.15.10